% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_clusterStat}
\alias{cytof_clusterStat}
\title{Statistics of the cluster results}
\usage{
cytof_clusterStat(
  data,
  markers,
  cluster = "cluster",
  sample,
  statMethod = c("mean", "median", "percentage", "NULL")
)
}
\arguments{
\item{data}{Input data frame.}

\item{markers}{The names of markers used for calcualtion.}

\item{cluster}{The column name contatining cluster labels.}

\item{sample}{The samples used for calculation.}

\item{statMethod}{Statistics containing mean, median or percentage.}
}
\value{
A matrix of the statistics results
}
\description{
Calculate the mean or median expression level of each marker for each cluster, or percentage
of cell numbers of each cluster for each sample.
}
