% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_heatmap}
\alias{cytof_heatmap}
\title{Heatmap plot of cluster mean value results}
\usage{
cytof_heatmap(
  data,
  baseName = "Cluster",
  scaleMethod = "none",
  dendrogram = c("both", "row", "column", "none"),
  colPalette = c("bluered", "greenred", "spectral1", "spectral2"),
  cex_row_label = NULL,
  cex_col_label = NULL,
  key.par = list(mgp = c(1.5, 0.5, 0), mar = c(3, 2.5, 3.5, 1)),
  keysize = 1.4,
  margins = c(2, 2)
)
}
\arguments{
\item{data}{A matrix with rownames and colnames}

\item{baseName}{The name as a prefix in the title of the heatmap.}

\item{scaleMethod}{Method indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is 'none'.}

\item{dendrogram}{Control the dengrogram on row or column, selection includes 'both', 'row', 'column', 'none'.}

\item{colPalette}{Use selected colour palette, includes 'bluered', 'greenred', 'spectral1' and 'spectral2'.}

\item{cex_row_label}{Text size for row labels.}

\item{cex_col_label}{Text size for column labels.}

\item{key.par}{Graphical parameters for the color key.}

\item{keysize}{Numeric value indicating the size of the key.}

\item{margins}{Numeric vector of length 2 containing the margins (see par(mar= *)) for column and row names, respectively.}
}
\value{
A heatmap object from \code{gplots}
}
\description{
Heatmap plot of cluster mean value results
}
