% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_progressionPlot}
\alias{cytof_progressionPlot}
\title{Progression plot}
\usage{
cytof_progressionPlot(
  data,
  markers,
  clusters,
  orderCol = "isomap_1",
  clusterCol = "cluster",
  reverseOrder = FALSE,
  addClusterLabel = TRUE,
  clusterLabelSize = 5,
  segmentSize = 0.5,
  min_expr = NULL,
  trend_formula = "expression ~ sm.ns(Pseudotime, df=3)"
)
}
\arguments{
\item{data}{The data frame for progression plot.}

\item{markers}{The column names of the selected markers for visualization.}

\item{clusters}{Select clusters for plotting, default selects all.}

\item{orderCol}{The column name of the estimated cell progression order.}

\item{clusterCol}{The column name of the cluster results.}

\item{reverseOrder}{If \verb{TRUE}, reverse the value of orderCol.}

\item{addClusterLabel}{If \verb{TRUE}, add the cluster label on the plot.}

\item{clusterLabelSize}{Size of the cluster label.}

\item{segmentSize}{Size of the cluster label arrow.}

\item{min_expr}{The threshold of the minimal expression value for markers.}

\item{trend_formula}{A symbolic description of the model to be fit.}
}
\value{
A ggplot2 object
}
\description{
Plot the expression trend along the estimated cell progressing order
}
