% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_shiny_dashboard.R
\name{cytofkit_shiny_dashboard}
\alias{cytofkit_shiny_dashboard}
\title{cytofkit_shiny_dashboard is used for performing the cytofkit2 analysis that include Data Pre-processing, Dimensionality Reduction, Clustering and Post-processing}
\usage{
cytofkit_shiny_dashboard(onServer = FALSE, port = NULL)
}
\arguments{
\item{onServer}{Whether the app is running on a server.}

\item{port}{Manually set a port, default is randomly allocated.}
}
\value{
Opens shinydashboard for cytofkit2 analysis
}
\description{
cytofkit_shiny_dashboard is used for performing the cytofkit2 analysis that include Data Pre-processing, Dimensionality Reduction, Clustering and Post-processing
}
