% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensVM.R
\name{DensVM}
\alias{DensVM}
\title{Density-based local maxima cluster with SVM}
\usage{
DensVM(ydata, xdata)
}
\arguments{
\item{ydata}{a matrix of the dimension reduced(transformed) data}

\item{xdata}{a matrix of the expression data}
}
\value{
a list contains a matrix \code{peakdata} of the peak numbers with different kernel bandwidth, and a matrix \code{clusters} of the cluster results
}
\description{
Density-based local maxima peak finding, subpopulation assigning with the power of SVM
}
\author{
Chen Jinmiao
}
