% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_shiny_dashboard.R
\name{cytofkit_shiny_dashboard}
\alias{cytofkit_shiny_dashboard}
\title{Title}
\usage{
cytofkit_shiny_dashboard(onServer = F, port = NULL)
}
\arguments{
\item{port}{Manually set a port, default is randomly allocated.}

\item{onSever}{Whether the app is running on a server.}
}
\value{

}
\description{
Title
}
